Lessening the Brotherhood's Technological Might!
This update will serve to slim down on some of the ammunition available at round-start to the Brotherhood, as-well as removing a layer of dense rock. The Brotherhood's bunker can be very difficult to penetrate - yet this caveat means for added creativity and planning may lead to more scenarios in which the Brotherhood isn't simply a funnel in form of fight.

This also buffs the health of a Ghoul Reaver safeguarding a semi-rich lootable area!